requirements.txt
torch~=1.3
torchvision~=0.4
matplotlib~=3.1
pillow~=6.2